fix(spool): bound output.log disk and memory growth (#405)
fix(spool): bound output.log disk and memory growth (#405)

5-layer fix for unbounded spool growth:

1. Legacy memory: bound output/stderr String accumulators with 2 MiB
   high-water drain (fixes 10+ GB OOM in legacy capture path)
2. SpoolRotator: file rotation at 32 MB with truncation sentinel
3. Consumer hardening: seek-based tail reads, UTF-8 lossy decode
4. Watchdog: monotonic bytes_written counter prevents false idle kills
5. Config: spool_max_mb (default 32) and spool_keep_rotated fields

Closes #405

+++ crates/csa-acp/src/client.rs
@@ -11,9 +11,11 @@ use agent_client_protocol::{
 /// Agent message and thought text beyond this limit is discarded from memory
 /// after being written to the output spool on disk.  1 MiB is sufficient for
 /// summary extraction and token-usage parsing, which only inspect the tail.
+///
+/// Canonical values shared with `csa-process::output_helpers` (same 1 MiB / 2 MiB).
 const TAIL_BUFFER_MAX_BYTES: usize = 1024 * 1024;
 
-/// High-water mark for tail buffer trimming (2x the target size).
+/// High-water mark for tail buffer trimming (2× the target size).
 ///
 /// We allow the buffer to grow to this size before trimming it back to
 /// [`TAIL_BUFFER_MAX_BYTES`].  This amortises the O(N) cost of
@@ -36,7 +38,7 @@ pub struct StreamingMetadata {
     /// Tail buffer of agent message/thought text (bounded by [`TAIL_BUFFER_MAX_BYTES`]).
     pub tail_text: String,
     /// Total bytes written to the output spool file.
-    pub(crate) spool_bytes_written: usize,
+    pub(crate) spool_bytes_written: u64,
 }
 
 impl StreamingMetadata {
